added util/mathematical_expr.h
Following the pattern in util/mathematical_types.h, expression classes that
don't exist immediately in usual programming languages are now in
util/mathematical_expr.h.
