+++ src/solvers/flattening/functions.h
@@ -14,7 +14,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <set>
 
-#include <util/std_expr.h>
+#include <util/mathematical_expr.h>
 
 #include <solvers/prop/prop_conv.h>
