+++ src/solvers/smt2/smt2_conv.cpp
@@ -20,6 +20,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <util/format_expr.h>
 #include <util/ieee_float.h>
 #include <util/invariant.h>
+#include <util/mathematical_expr.h>
 #include <util/pointer_offset_size.h>
 #include <util/std_expr.h>
 #include <util/std_types.h>
