+++ src/solvers/smt2/smt2_parser.cpp
@@ -13,6 +13,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <util/arith_tools.h>
 #include <util/ieee_float.h>
 #include <util/invariant.h>
+#include <util/mathematical_expr.h>
 #include <util/range.h>
 
 #include <numeric>
