+++ src/util/format_expr.cpp
@@ -19,6 +19,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "format_type.h"
 #include "ieee_float.h"
 #include "invariant.h"
+#include "mathematical_expr.h"
 #include "mp_arith.h"
 #include "rational.h"
 #include "rational_tools.h"
