+++ src/util/simplify_expr_boolean.cpp
@@ -13,6 +13,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include "expr.h"
 #include "expr_util.h"
 #include "invariant.h"
+#include "mathematical_expr.h"
 #include "namespace.h"
 #include "std_expr.h"
